Do not delete depscan and envupdate caches on reboot, as this should improve bootup times - works nicely with the '-u' parameters of both. Also move the '-u' parameter test of depscan.sh after the directory exist test, else /started/, etc do not exist.
svn path=/trunk/; revision=652

+++ sbin/depscan.sh
@@ -5,25 +5,6 @@
 
 source /etc/init.d/functions.sh
 
-# Only update if files have actually changed
-update=1
-if [ "$1" == "-u" ]
-then
-	update=0
-	for config in /etc/conf.d /etc/init.d /etc/rc.conf
-	do
-		if [ "${config}" -nt "${svcdir}/depcache" ]
-		then
-			update=1
-			break
-		fi
-	done
-	shift
-fi
-[ ${update} -eq 0 ] && exit 0
-
-ebegin "Caching service dependencies"
-
 if [ ! -d "${svcdir}" ]
 then
 	if ! mkdir -p -m 0755 "${svcdir}" 2>/dev/null
@@ -42,6 +23,25 @@ do
 		fi
 	fi
 done
+
+# Only update if files have actually changed
+update=1
+if [ "$1" == "-u" ]
+then
+	update=0
+	for config in /etc/conf.d /etc/init.d /etc/rc.conf
+	do
+		if [ "${config}" -nt "${svcdir}/depcache" ]
+		then
+			update=1
+			break
+		fi
+	done
+	shift
+fi
+[ ${update} -eq 0 ] && exit 0
+
+ebegin "Caching service dependencies"
 
 # Clean out the non volitile directories ...
 rm -rf "${svcdir}"/dep{cache,tree} "${svcdir}"/{broken,snapshot}/*
